requirements.txt
empath==0.89
gensim==4.3.2
httpx==0.25.0
ipywidgets==8.1.1
jupyterlab==4.0.7
matplotlib==3.8.0
pandas==2.1.2
pillow==10.1.0
pyalex==0.12
pybliometrics==3.5.2
rich==13.6.0
scikit-learn==1.3.2
scipy==1.11.3
spacy==3.7.2
seaborn==0.13.0
wordcloud==1.9.2